requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
Django==1.11.17
django-admin-sortable2==0.6.19
django-classy-tags==0.8.0
django-cms==3.5.3
django-debug-toolbar==1.9.1
django-el-pagination==3.2.4
django-filer==1.3.0
django-formtools==2.1
django-google-tag-manager==0.0.3
django-mailer==1.2.2
django-modeltranslation==0.12.2
django-mptt==0.8.7
django-polymorphic==1.3
django-rosetta==0.7.14
django-sekizai==0.10.0
django-treebeard==4.2.1
django-webpack-loader==0.5.0
easy-thumbnails==2.4.2
html5lib==0.9999999
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
lockfile==0.12.2
mccabe==0.6.1
microsofttranslator==0.8
Pillow==5.1.0
polib==1.1.0
psycopg2==2.7.3.2
pylint==1.7.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
python-memcached==1.58
pytz==2018.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
wrapt==1.10.11
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.2.1
djangocms-picture==2.0.6
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
django-geoposition-2==0.3.6
django-select2==6.3.1
aldryn-apphooks-config==0.4.1
